change dashbord functionality

+++ database/migrations/2026_04_24_095125_create_user_doctor_role_membership_table.php
@@ -4,20 +4,20 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
-        Schema::create('user_doctor_role_membership', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->decimal('membership_amount', 10, 2)->default(0);
-            $table->date('membership_subscription_date')->nullable();
-            $table->date('membership_subscription_end_date')->nullable();
-            $table->timestamps();
-
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('user_doctor_role_membership')) {
+            Schema::create('user_doctor_role_membership', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->decimal('membership_amount', 10, 2)->default(0);
+                $table->date('membership_subscription_date')->nullable();
+                $table->date('membership_subscription_end_date')->nullable();
+                $table->timestamps();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
